moving the macro commands expr_not to boolean
